Create FindMinXOR.java
You are given an array A of size N which needs to divide the array into one or more contiguous intervals. 
For each interval created. It needs to calculate the bitwise OR of all the elements in that interval and 
then determine the bitwise XOR of all the values she gets by doing the OR of the respective intervals.
your task is to find and return an integer value representing the minimum possible XOR value that can achieve